Declare integer locals as 64-bit
The local variables were declared with the type inferred from their
initial value, so that ``result = 0`` became an ``int`` in C#
(``var``), Java (``var``), Go (``:=``) and C++ (``auto``), while we
represent the integers of the meta-model with 64 bits. A subsequent
assignment of an integer argument, *e.g.*, in a branch of a switch,
therefore failed to compile in C#, Java and Go, and silently narrowed
the value in C++.

We now declare the integer locals as ``long`` in C# and Java, and
convert the integer literals to ``int64`` in Go. In C++, we spell out
all the primitive types of the locals instead of ``auto``, which also
avoids deducing ``const wchar_t*`` for the string literals.

Along the same lines, ``str.find`` gave an ``int`` in C# and Java, both
from the native ``IndexOf`` and ``indexOf`` and from the helpers. We now
return ``long`` from the helpers and cast the native positions, as
a boxed ``Integer`` never equals a boxed ``Long`` in Java.

The common meta-model ``switch_statements`` gets a verification function
which re-assigns an integer local in a switch branch.

+++ dev/test_data/main/java/expected/string_slicing_and_find/expected_output/src/main/java/dummy/common/StringHelpers.java
@@ -83,7 +83,7 @@ public final class StringHelpers {
    * @param start of the search
    * @return the position of {@code sub} in {@code text}, or -1 if not found
    */
-  public static int find(String text, String sub, long start) {
+  public static long find(String text, String sub, long start) {
     final long theStart = start < 0 ? Math.max(start + text.length(), 0) : start;
     if (theStart > text.length()) {
       return -1;
